+++ CHANGELOG.md
@@ -4,17 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.9.0] - [2020-04-08]
-### Changed
+## [0.10.0] - [2020-04-08]
+### Added
 - Add Trading Partner API querying capability
 - Request::TradingPartner
 - Response::TradingPartnerData
 - Response::TradingPartnersData
 - Models::TradingPartner
 
+## [0.9.0] - [2020-04-08]
+### Added
+- Added EligibilityBenefit#medicare?
+- Added EligibilityData#medicare?
+
+### Changed
+- EligibilityBenefit(s)#individual returns true for medicare in benefit and queries that look for 'IND'
+- EligibilityBenefit(s)#in\_network returns true for medicare in benefit and queries that look for 'Y'
+
 ## [0.8.0] - [2020-04-04]
 ### Changed
-- Attemp to serialize all properties with 'date' in the name to ChangeHealth date format
+- Attempt to serialize all properties with 'date' in the name to ChangeHealth date format
 
 ## [0.7.0] - [2020-04-03]
 ### Changed
@@ -65,7 +74,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Authentication
 - Configuration
 
-[0.9.0]: https://github.com/WeInfuse/change_health/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/WeInfuse/change_health/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/WeInfuse/change_health/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/WeInfuse/change_health/compare/v0.5.0...v0.6.0